backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
pandas==2.2.3
scikit-learn==1.6.0
numpy==2.2.1
matplotlib==3.10.0
joblib==1.4.2
requests==2.32.3